Add code to collect and prepare forecast data.
Generate forecast data by copying point and interval data
into an own timeslice. For this we're using intervals that
are generally available for all days, but in case some don't
exist, we'll search back and forward in an alternating
pattern until a limit is reached and try to get the most
useful data.