Support for night icons (current weather and forecast).
Since the migration to the new data provider, the different
appearance between night and day forecasts has been lost.
Depending on your location, it might be a bit strange to
have sunny nights. This patch adds back support for night
and day icons.

There are only a couple of icons which are different at night.
Let the calling functions find out if it's night or day, and
get_icon can then choose the right icon.