Use a shortened location name in the summary window.
When the location name is too big it causes the summary window
to grow, making the layout look awful. So for the subtitle, shorten
the location name and only use the name of the city.

We might want to use the shortened name elsewhere, and maybe
let the user change it, so let's serialize it.