Locale support, part 1: Duplicate strings.
Values like temperature, wind speed etc. are returned as immutable
strings. That's bad because those values are not printed in the
format defined by the user's locale.

As a first step, duplicate the strings in get_data() and free them
in the calling functions. Remove usage of CHK_NULL from get_unit(),
as it would create a memory leak.