Rework code to get current weather.
For future plans we need to rework the code which gets data about
the current weather. We do this in a manner similar to getting
data for a forecast, only this time we search for the shortest
interval possible, preferring a timeslice near the current time.

While this commit won't break compilation, it introduces
memory leaks which will be solved in subsequent commits.