Use extra field in xml_time for the time of point data.
Enhance xml_time with an extra field holding the time of
point data. While this provides extra info for the user, its main
intention is for making debugging a bit easier, in case something
goes wrong or the data provided changes in a way we did not
expect.

Also, we do not mix up times of point and interval data in the
forecast anymore, which was only a transitional solution.

In the forecast details tab, we can now present the times for
both point and interval data. While doing so, add a short
explanation what these times are about.