Use a consistent naming scheme for time_t and struct tm variables.
As far as consistent naming is concerned, the current code is really
a mess. Sometimes start_t is a time_t, sometimes it's a tm struct,
sometimes start_ts is used as a name...
So let's cleanup and put an end to this time_t/tm hell once and
for all.