Try to make the forecast scrolled window a bit nicer.
Relying on the automatic gtk widget size management, I was terribly
unsuccessful producing a nice result. So in a new attempt, let's
try to do the problematic stuff manually:

* Use a fixed size of 150px for the forecast boxes.
  That makes the table look a bit more homogenous.

* Use a fixed width for the scrolled window:
  - 650 pixel for monitor width <= 720
  - 700 pixel for anything bigger

* Stuff the scrolled window into an event box to make it look more
  consistent with the scroll-less version.

Because of the fixed width, some labels in the forecast boxes can
now be cut off. If that's the case, we'll need to tweak/enlarge
the forecast box width, it can't be that much off. I hope the
current setting works for most translations, though.