Do away with the short location name.
Since the forecast data is requested using latitude and longitude,
we're free to set the location name to something arbitrary. As a
first step, sanitize the location name returned by the search
dialog or by auto-detection via IP address and remove the code added
previously for keeping an extra short location name.