Really get the plugin working on OpenBSD (bug #9152).
The #define added in 94c7c6e88ed7caf1ed8fec608558050274064a84
is only part of the solution. One also has to initialize the
tm structs before using strptime, otherwise the behaviour is
not defined.

Some references:
http://linux.die.net/man/3/strptime
"In principle, this function does not initialize tm but only stores the values
specified. This means that tm should be initialized before the call. Details
differ a bit between different UNIX systems. The glibc implementation does not
touch those fields which are not explicitly specified, except that it
recomputes the tm_wday and tm_yday field if any of the year, month, or day
elements changed."

http://www.openbsd.org/cgi-bin/man.cgi?query=strptime
"There is no way to specify whether Daylight Saving Time is in effect when
calling strptime.  To use the resulting tm structure with functions that
check the tm_isdst field, either set it to a negative value, which will
cause mktime(3) to attempt to divine whether Daylight Saving Time would
be in effect for the given time, or compute the value manually."