Astrological data: Use data for day/night determination.
Finally we can use sunrise and sunset times for determining
whether it is (polar) day or (polar) night. For that, add a
new field to xfceweather_data which can be used in most parts
of the code.

This also brings along a rework of the several update functions,
some smaller fixes and reformatting.