Replace my_timegm() with something that doesn't lock up.
It seems the original my_timegm from the Linux manpage that is
supposed to be a portable replacement of timegm() does not work
very well when we use multiple threads, such as when parsing
astrological data and weather data at the same time.
On occasion, the weather plugin would lock up somewhere in
tzset(). With this new version, this does not happen anymore.

It does do calculations with time_t however, so it may not be
as portable as it could be.