Clear scrollbox labels before updating conditions (bug #9210).
When updating current conditions (that is, switching to a new data
point in already downloaded data), currently shown labels are not
cleared before adding labels referring to the new conditions.

Result: There are two temperature labels displayed, with animation
between them. Both show the same value. After another hour there
are three labels - two with the old value and one with the new.