Refactor icon and scrollbox update code.
set_icon_current() and set_icon_error() do more than just updating
the icon in the panel. set_icon_error() frees weather data, for
example, which is not good and may cause nasty surprises, and both
update the scrollbox, which might not be desired.

This commit splits up the updating code into separate functions
for the icon and the scrollbox. These take care of properly setting
the icon and labels according to the data available, so they are
only responsible for the representation, but will not change
weather data, which should be done elsewhere.