Initial support for custom icon themes.
This adds rudimentary support for loading custom icon themes.
GUI code for easy selection is still missing, but one can
edit the panel plugin rc file and add a theme_dir entry pointing
to the (absolute) path of the custom theme.

The liquid icon theme's directory structure has been reorganized
to reflect the changes and shall serve as a model for new themes.