Prepare usage of libsoup.
The current implementation of weather-http.c would need to be modified
to support HTTP/1.1 chunked transfer encoding. While implementing the
appropriate decoding algorithm is certainly possible, the better idea
is to rely on and use an existing implementation. This will make many
things much easier and faster to achieve and solve any outstanding
bugs that are still present in the current implementation.

libsoup should be available on any machine where Xfce is installed,
so this additional dependency should not be a problem. The screenshot
plugin is already using it too.