Implement multi-lines for the scrollbox (bug #5722).
It works fine, except that there is no easy way to align the pango layout
vertically. This will only be noticeable when there are fewer values than
lines, so I consider this only a minor problem which can be solved with
proper configuration (change number of lines or add values to fill up the
empty lines).