Schedule data update after location parameter changes.
When the user changes any of the lat/lon/alt settings, then new weather
and astro data need to be downloaded. Of course, the user needs some time
to make changes, so an update should not occur too often because that could
send many requests to the webservices. To minimize these requests, install
a timeout that gets reset whenever the user changes these settings.

Note that the main update routine has an update timeout too that may add
to this timeout, so the download will not happen immediately after this
timeout expires.