Implement data interpolation for current weather.
If enough point data is available, we can interpolate data between start
and end of the interval. It seems to work fine, so let's remove the config
option and do it by default, computing current conditions every five minutes.

Default back to choosing point data at the end of the interval in case
interpolation is not possible.

Some values need to get special treatment though, which will be dealt with
future commits.