Set expiration date on cache file.
Primarily the purpose of the cache file is to preserve data across panel
restarts. It won't make much sense to keep old forecasts, they will be
wrong anyway (if they were ever right). So better put an expiration
date on the cache file and do not rely on old data. Let's try 2 days
first - though I'd go for 24 hours actually - and tune it later, or
perhaps make it a (hidden?) option.