Fix crashes related to updates and timer handling.
Closing the dialog will clear its data, so the update routine will only get
a NULL pointer and cause a crash. Reading data from the cache file is not
the task of the config dialog, but of the update_with_reset function, so
it belongs there and read_cache_file can be static. Also it shouldn't happen
after we have downloaded the new data but before it, so the new data won't
get overwritten by the old data. Finally, set timer ids to 0 so the checks
for them are actually useful.