Modify the symbol according to calculated point data.
Since we're interpolating data for both current conditions and
forecast, let's adapt the symbol a bit to better match the values
at the time chosen. This way the values make more sense, and
to keep as close as possible to the original data, only the
amount of clouds (lightly cloudy, partly cloudy, cloudy) is
changed.