Time for more fog.
It seems met.no doesn't use the FOG symbol at all?
Let's try to safely make use of it by analysing the fog value
and the other values, and choose it if it seems right to do so
(it's still low priority, but not that much low priority as
before).