Make timezone a double value and integrate dst value into it.
There are timezones/dsts that are just half an hour off, not a whole hour,
e.g. Australian locations.

Let's include daylight saving times into the timezone value and increase
limits to +/- 25 hours. Without adding additional dependencies, the
user has to alter the correction manually. This will be improved in
the next version (glib version has to be increased to 2.26 for GDateTime
and GTimezone support).