New value: Apparent temperature (bug #9564).
Apparent temperature, also known as "felt temperature" or "feels like" value
is a debatable and rather subjective value, as can be deduced from the
various models available that try to compete for correctness and accuracy.

The different models that are offered by the plugin are collected from
various sources and/or derived from publicly available specifications,
the sources are mentioned in the code. Many commercial weather data
providers use their own proprietary (sometimes patented) models.

* Windchill/Heat index:
  Used in North America, wind chill will be reported for low temperatures
  and heat index for higher ones. At night, heat index will be replaced by
  the Summer Simmer Index.

* Windchill/Humidex:
  The Canadian counterpart, with the wind chill being similar to the
  previous model but with slightly different constraints. Instead of
  the heat index Humidex will be used.

* Steadman:
  This is the model used in Australia, especially adapted for the climate
  of this continent. Maybe in Central Europe too, but then windchill and
  similar values had never gained so much popularity as in the US or Canada,
  so information about its usage is scarce. It can be used for lower and
  higher temperatures, so no extra heat index is required.

* Quayle-Steadman:
  Improvements made to earlier experiments/developments by Steadman in
  1998.

These models are not better or worse than others; Usually it's advised to
use the model for your country as the formulas have been adapted to account
for the local climate and resemble what you get presented in the news.
The main reason for the existence of these models is being able to warn
about the risk of frostbite and heat strokes. Otherwise, apparent
temperature is a rather subjective value and both reliability and
significance are not very high.