Forecast tab: Handle missing values (humidity, fog and cloud data).
Some values are missing in forecasts more than 3 days into the future:
Humidity, fog, cloud data. Further, because humidity is required for
the calculations, the dew point cannot be computed. This patch handles
this case and thus fixes the tooltips of the affected forecast cells.

Example location for reference: Oslo, Norway.