Use panel toggle button instead of event box as the main plugin container.
This replaces the event box (called tooltipbox here) with a panel toggle
button which other plugins use, e.g. the datetime plugin. This toggle button
provides better feedback for the user: It indicates that something happens
when the user clicks on it, and the active (pressed) state of the button
shows that the forecast window is open (useful in case that it is hidden
somewhere below the other windows or on another workspace).

While we're at it, let's finally replace top_hbox and top_vbox with alignbox,
which is an xfce_hvbox and more convenient for handling panel orientation
changes.