Synchronise the toggle button state with the summary window.
This needs a bit of trickery, because the toggle signal is emitted whenever
we want to change its state. So we will handle the toggle signal itself
in a callback function which sets the proper state while blocking another
toggle signal. The actual clicks are still handled by cb_click.