Summary window: Allow aligning met.no image to the left for RTL.
RTL languages like arabic or hebrew encounter a problem when translating
the details tab: Since the image is aligned to the right, it would be
overwritten by the text, or the text would be written over the image.
This patch gives the translator the possibility to align the image to the
left by providing a translatable string that is checked at runtime. If
it is the default (LTR), then the image will be aligned to the right,
otherwise it will be aligned to the left.

Non-RTL translators can simply leave this string untranslated, or set
it to "LTR".