Modify symbol only for current conditions.
Except for fog, leave the symbols on the forecast tab untouched and stay
true to the original data. The difference in interval data and point data
should be clear to the user. Only for the "current weather" it makes sense
to change the symbol, and maybe more work on calculating the symbol will
be spent for future versions.