Summary window: Optimize search for daytime intervals.
Choose only the 6 hours intervals that start at 0, 6, 12, 18 hours
UTC time, not the first best 6 hours interval found.

This makes the results similar to those presented on the yr.no website,
except we try to show raw data at 9, 15, 21, 3 o'clock ±1 hour.

Also show the interval for current conditions if a 6 hours daytime
interval is not available.