Fix refresh function.
The current implementation of that function with asynchronous downloads
is a bit weird. When the user hits refresh, the plugin should do exactly
that, not show cached data which may be replaced by downloaded data some
time later. It's also a bit unorthodox to require a "clear" parameter
in a reset function, so let's clean up this mess.

This commit sets the following new behaviour:
If the user hits refresh, then "No data" is displayed at once, all data is
cleared, cached data read and downloads are scheduled immediately (and for
real). If the download finishes before the standard update interval (10
secs) runs the GUI update, then new data will be shown, if not then cached
data will be shown until download is finished. If the download fails and
cached data is available, cached data will be shown immediately.

The scrollbox and icons are now reset synchronously and immediately after a
download.