Better handle single row and icon size in various panel modes.
Add an option to set the single row property manually, defaulting to TRUE.
In deskbar mode, enabling this means that the icon will always be in the same
"row" as the scrollbox, and never get bigger than one row.

If there are more than two rows, the icon will be scaled to 80%. This is
just an arbitrary value, but most icons tend to look better if they do not
fill the whole panel at bigger sizes.