Fetch and cache astronomical data for multiple days.
met.no sunrise API provides astronomical data for up to 30 days in the
future (requesting more than that will result in an error page).
We can make use of that and cache it, similar as we do for the forecast
data. current_astro will point to the astronomical data for the current
day.

Download will now happen in intervals of 24 hours, not at a fixed time
at midnight as previously. The current conditions updates will take
care that astronomical data gets updated for the present day, because
they happen exactly at midnight and thus save us from adding additional
logic having to deal with that.