Somewhat fix the update handling.
The recent changes have broken the update code. To be honest, the current
way of dealing with updates is not very nice and probably calls for a
redesign.

This patch cleans up the mess a bit. While it (hopefully) fixes update
handling functionally, conceptually it leaves it broken. However, being
usable and with the changes that are planned for the near future in mind,
time is better spent working on a better solution that fulfills the new
requirements than trying too hard to fix it now.