Make plugin ready for met.no locationforecast-1.2 API (bug #10916).
http://api.yr.no/weatherapi/locationforecastlts/1.1/documentation#version_1_2___2014_05_20

The updated API version uses CamelCase symbol names instead of UPPERCASE
ones and has added some new names (like 'Drizzle'), which unfortunately
have not been documented (yet?).

What's more, the typo 'celcius' has been replaced by the fixed 'celsius',
but luckily the plugin will not be affected by that change.

What's a bit more unfortunate is that the existing icon themes do not
cover the new symbol names, so one would have to create quite a lot of
new icons. Also, new translations would have to be added for the symbols,
and maybe existing ones would have to be altered. To prevent this, we're
simply going to map the new symbols to existing ones as good as possible.
This should be good enough for the time being.