Fix processing day of astrodata.
The way the day information of astrodata is processed is erroneous
and leads to different problems regarding localtime and timezones.
This commit fixes this by extending parse_timestring to generate
either local or gmt times and make the other parts use this for
parsing astrodata instead of applying clumsy, non-functional
workarounds.